requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
Pillow>=9.0.0
python-dotenv>=0.19.0
requests>=2.28.0
transformers>=4.21.0
torch>=1.12.0